demo1/pyproject.toml
[project]
name = "demo1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-generativeai>=0.8.5",
    "pandas>=2.3.2",
    "phidata>=2.7.10",
    "pinecone==7.0.0",
    "pypdf2>=3.0.0",
    "pyreadline3>=3.5.4",
    "python-dotenv>=1.1.1",
    "sentence-transformers>=5.1.1",
    "streamlit>=1.50.0",
]
